requirement.txt
streamlit
google-generativeai
pandas
plotly
fpdf
python-dotenv